Refractored actionURL and move the function into compute/v2/extensions
