--- openstack/compute/v2/extensions/evacuate/urls.go
@@ -1,9 +1,0 @@
-package evacuate
-
-import (
-	"github.com/gophercloud/gophercloud"
-)
-
-func actionURL(client *gophercloud.ServiceClient, id string) string {
-	return client.ServiceURL("servers", id, "action")
-}
