+++ openstack/compute/v2/extensions/resetstate/requests.go
@@ -2,6 +2,7 @@ package resetstate
 
 import (
 	"github.com/gophercloud/gophercloud"
+	"github.com/gophercloud/gophercloud/openstack/compute/v2/extensions"
 )
 
 // ServerState refers to the states usable in ResetState Action
@@ -18,6 +19,6 @@ const (
 // ResetState will reset the state of a server
 func ResetState(client *gophercloud.ServiceClient, id string, state ServerState) (r ResetResult) {
 	stateMap := map[string]interface{}{"state": state}
-	_, r.Err = client.Post(actionURL(client, id), map[string]interface{}{"os-resetState": stateMap}, nil, nil)
+	_, r.Err = client.Post(extensions.ActionURL(client, id), map[string]interface{}{"os-resetState": stateMap}, nil, nil)
 	return
 }
