--- openstack/compute/v2/extensions/resetstate/urls.go
@@ -1,9 +1,0 @@
-package resetstate
-
-import (
-	"github.com/gophercloud/gophercloud"
-)
-
-func actionURL(client *gophercloud.ServiceClient, id string) string {
-	return client.ServiceURL("servers", id, "action")
-}
