+++ openstack/compute/v2/extensions/urls.go
@@ -1,7 +1,7 @@
-package rescueunrescue
+package extensions
 
 import "github.com/gophercloud/gophercloud"
 
-func actionURL(client *gophercloud.ServiceClient, id string) string {
+func ActionURL(client *gophercloud.ServiceClient, id string) string {
 	return client.ServiceURL("servers", id, "action")
 }
